api/requirements.txt
annotated-types==0.7.0
anyio==4.5.0
asgiref==3.8.1
CacheControl==0.14.0
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
cryptography==43.0.1
distro==1.9.0
Django==5.0
django-cors-headers==4.3.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
drf-nested-routers==0.94.1
firebase-admin==6.5.0
google-ai-generativelanguage==0.6.9
google-api-core==2.19.2
google-auth==2.34.0
google-auth-httplib2==0.2.0
google-cloud-core==2.4.1
google-cloud-firestore==2.18.0
google-cloud-storage==2.18.2
google-crc32c==1.6.0
google-generativeai==0.8.1
google-resumable-media==2.7.2
googleapis-common-protos==1.65.0
grpcio==1.66.1
grpcio-status==1.66.1
h11==0.14.0
httpcore==1.0.5
httplib2==0.22.0
httpx==0.27.2
idna==3.10
jiter==0.5.0
msgpack==1.1.0
openai==1.46.1
proto-plus==1.24.0
protobuf==5.28.2
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
PyJWT==2.8.0
pyparsing==3.1.4
python-dotenv==1.0.1
pytz==2024.1
requests==2.32.3
rsa==4.9
sniffio==1.3.1
sqlparse==0.5.0
tqdm==4.66.5
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.3
psycopg2-binary==2.9.10
pandas==2.1.4